Re-write processed text (CoNLL fmt) as one sentence per line text (#2)
This PR adds a new program, `conll-to-text-line`, which transforms a processed text file, in CoNLL format, as a simple, unprocessed, text file. The output format is a sentence on each line. Care is taken during the transformation to merge tokens appropriately (e.g. `["do", "n't"] => "don't"`) and do individual token transformations (e.g. `"-LRB-" => "("`) as well as proper token spacing (e.g. no space between a word and a period).

Currently, the original sentence words are output. It would be relatively easy to have the output tokens be lemmatized instead.